chore: Move to utils instead of constants
