+++ frontend/src/scenes/hog-functions/list/hogFunctionsListLogic.tsx
@@ -7,7 +7,8 @@ import { lemonToast } from '@posthog/lemon-ui'
 
 import api from 'lib/api'
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
-import { buildAlertFilterConfig, objectsEqual } from 'lib/utils'
+import { objectsEqual } from 'lib/utils'
+import { buildAlertFilterConfig } from 'lib/utils/alertUtils'
 import { deleteWithUndo } from 'lib/utils/deleteWithUndo'
 import { projectLogic } from 'scenes/projectLogic'
 import { userLogic } from 'scenes/userLogic'
